feat(model+data): variable scale factor + DLSS-style presets + Halton jitter
Three pre-T5-launch wins:

1. ORUPico.scale_factor is now a constructor arg (default 2.0). Verified
   working at 1.0 (DLAA), 1.3, 1.5, 1.7, 2.0, 2.5, 3.0, 3.3 — any positive
   float. Previously hardcoded to 2.0.
2. PRESET_SCALE_FACTORS + preset_scale() helper: DLAA / Ultra Quality /
   Quality / Balanced / Performance / Ultra Performance, mirroring DLSS
   convention. Case-insensitive + alias-friendly ('q', 'ultra-perf', etc).
3. NoiseBase loader synthesizes per-frame Halton(2,3) sub-pixel jitter
   when deriving LR from HR. Critical: without jitter, the recurrent state
   can't learn temporal sub-pixel accumulation — model degenerates to
   single-frame SR. With jitter, consecutive frames sample HR at slightly
   shifted positions, motion vectors carry the jitter delta, network
   learns proper temporal SR (same trick FSR2/DLSS use).

Tests: 83 passed, 3 skipped (unchanged).

+++ ors/model/wavelet.py
@@ -258,7 +258,7 @@ class WaveletKPNHead(nn.Module):
         self,
         feature_ch: int = 32,
         kernel_size: int = 5,
-        scale_factor: int = 2,
+        scale_factor: float = 2.0,
         levels: int = 2,
         wavelet: str = "db2",
     ):
@@ -266,7 +266,7 @@ class WaveletKPNHead(nn.Module):
         self.feature_ch = feature_ch
         self.k = kernel_size
         self.k2 = kernel_size * kernel_size
-        self.scale_factor = scale_factor
+        self.scale_factor = float(scale_factor)
         self.levels = levels
 
         self.swt = SWT2D(levels=levels, wavelet=wavelet)
